Update: PPInclude.

+++ .vitepress/config/zh.ts
@@ -34,7 +34,7 @@ export const zh = defineConfig({
         lastUpdated: {
             text: '最后更新于',
             formatOptions: {
-                dateStyle: 'short',
+                dateStyle: 'long',
                 timeStyle: 'medium'
             }
         },
